+++ src/main/java/com/titancake/TitanCake/repository/RecetaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.titancake.TitanCake.model.Receta;
 
-public interface RecetaRepository extends JpaRepository<Receta, Long>  {
+public interface RecetaRepository extends JpaRepository<Receta, Integer>  {
     
 }
